requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.30.6
jinja2==3.1.5
pydantic==2.10.6
python-dotenv==1.0.1
openai>=1.0.0
pypdf==5.3.0
